Rename `matrix::MatrixConfig` to `matrix::Config`
Sorry, Yassine... But I *really* don't like repeating the module name in the type name.
It somewhat defeats the purpose of having modules, and it's a naming convention that is
not used anywhere else in the code, including by third-party libraries. The Matrix client
is at `matrix_sdk::Client`, not `matrix_sdk::MatrixClient', etc.

Would like to maintain consistency.

I alias it to MatrixConfig in `main`.