WIP: modularize listener capabilities
before, we had everything crumbled on a single file `src/api.rs` which
was ugly, and verrry messy, and this is an attempt to get over that by
separating distinctive modules, mainly, ws, node, redis, mail and matrix
modules.

This is still a WIP, and things my change dramatically in the future
from either design choices or adapting to upstream changes.