fix(#73): add connection timeout for redis
Before, we had a spawned tokio task in `spawn_redis_listener` that
listens for redis events using event stream without handing the case
of a dead stream (stream is valid but no more events are expected),
this approach resulted in spawning multiple "dead" connections.

This commit introduces a listener timeout, that controls for how long a
listener should wait for a given, singular redis event (re-new each time
we receieve an event), this duration can be configured in the redis
config filds under the name `listener_timeout` with duration represented
in seconds.