feat: index old registrations in postgres DB (#70)
* feat: index old registrations in postgres DB

Iterate over old chain blocks, starting from block 0 to index old
registrations in postgres db, the state of the indexer (current blocks
number and hash) is also saved in db to prevent re-iterating each time we
start the registrar backend.

* refactor: use `IdentityOf` to fetch identities

Instead of iterating throught the blocks starting by the first we now:
- Fetch all idnetities up to the most recent block if no `indexer_state`
  is present
- Iterate over the missing blocks starting from the marked block in the
  `indexer_state` to the most recent block checking for `JudgementGiven`
  events with

Also, only we only register identities with `Reasoable` judgement

* fix: index new registration in case of conflict

Before if we have an indexed row in db, and try to overwrite/update
that row with new data, we get an error since we use an INSERT query
for but creating and updating, now we handle that case buy adding an
`ON CONFLICT` statement at the end favoring the newly added data to
overwrite the old ones.

* fix: index missing identities by "diffing"

Before, we used to index missing identities (that are not yet indexed in
our db) by iterating over missing blocks one by one and checking for
JudgementGiven event, which takes allot of compute power and time, now
we fetch all on chain and on db identities, diff them, then insert the
missing identities to db.