Fix/major fixes (#83)
* fix: update postgres EVENT type

Since the postgres EVENT enum and the (Rust) local Event enum was
different after the introduction of the image varient in Rust, causing
new registration to halt after initial `requestJudgement` request
(challenge updates fail), this commit act as a fix to that issue by
adding the `image` varient to the postgres EVENT type

* fix: properly pass NETWORK in postgres select query

Includes other minor changes, mainly formatting.

* fix: update README with all possible WS request types

* feat(braking): reform pending_challenges returned field

`pending_challenges` field previously is an Vec of Vecs that has
the account type and challenge as a two fields in the inner Vec, but
now it is a Vec of Challenges (json obj:s) that contain three fields,
`challenge`, `account_name` and `account_type`.
{"account_name": [name], "account_type": [type], "challenge": [challenge]}

WARNING: THIS COMMIT INTRODUCES BREAKING CHANGE THAT IS NOT COMPATIBLE WITH
OLDER COMMITS.

* fix: port breaking changes to API 1.1 and keep legacy in 1.0

* Changes
- Fix typos (as usual)
- Port breaking changes to newer API version `1.1`
- Keep legacy API version `1.0` for compatibility reasons
- Update Readme with the new API version