feat: implement google like search behavior
Descriptoin:
Added a new search field named `Generic` that will act as a generic
(obviously) search for all fields AKA fuzzy finder, how this work is by
having a new column in the registration field named `search_text` that
contain the concatinated version of all the previous columns (auto
update itself on insertion) and then we use the postgres `similarity`
function to fzf/search for rows and rank them by how similar they are.

Note:
This can be chained with other search fields like `Discord`, `Matrix`,
etc. And in that case, they will act as a chain of filters, and return
rows that satisfies ALL matching requirements.