add optimistic code generation
this is now very insecure way to do things, but
i dont want to compile on my laptop, it is what it is.
DANGER: ddoss anyone could spam  endless amount of entries with
websocket conncetions and it might even trigger us to automaticly
send outgoing verifications so yeah. TODO FIXFIX WHEN NOT IN NIXNIX